requirements.txt
Pillow>=10.0
pillow-avif-plugin>=1.5
piexif>=1.1.3
click>=8.1
customtkinter>=5.2
